refactor : 통계와 검색 책임 분리
