+++ srcs/problem/problem.model.js
@@ -2,74 +2,6 @@ import { pool } from "../../config/db.js";
 import { sql } from "./problem.sql.js"; 
 
 export const ProblemModel = {
-
-  searchAll: async (query) => {
-    try {
-      console.log("searchAll");
-      const [results] = await pool.query(sql.searchProblem, [`%${query}%`]);
-      return results;
-    } catch (error) {
-      console.error("문제 검색 실패:", error.message);
-      throw new Error("문제 검색 실패");
-    }
-  },
-
-  searchByFolder: async (query, folderId) => {
-    try {
-      const [results] = await pool.query(sql.searchProblemByFolder, [`%${query}%`, folderId]);
-      return results;
-    } catch (error) {
-      console.error("문제 검색 실패:", error.message);
-      throw new Error("문제 검색 실패");
-    }
-  },
-
-  checkSubscriptionStatus: async (userId) => {
-    try {
-      const [results] = await pool.query(sql.checkSubscriptionStatus, [userId]);
-      return results.length > 0;
-    } catch (error) {
-      console.error("구독 상태 확인 실패:", error.message);
-      throw new Error("구독 상태 확인 실패");
-    }
-  },
-
-
-
-  findById: async (problemId) => {
-    try {
-      const [results] = await pool.query(sql.findProblemById, [problemId]);
-      if (results.length === 0) {
-        return null;
-      }
-      return results[0];
-    } catch (error) {
-      console.error("문제 조회 실패:", error.message);
-      throw new Error("문제 조회 실패");
-    }
-  },
-
-  findPhotosByProblemId: async (problemId) => {
-    try {
-      const [results] = await pool.query(sql.findPhotosByProblemId, [problemId]);
-      return results;
-    } catch (error) {
-      console.error("사진 조회 실패:", error.message);
-      throw new Error("사진 조회 실패");
-    }
-  },
-
-
-  findTypesByProblemId: async (problemId) => {
-    try {
-      const [results] = await pool.query(sql.findTypesByProblemId, [problemId]);
-      return results;
-    } catch (error) {
-      console.error("유형 조회 실패:", error.message);
-      throw new Error("유형 조회 실패");
-    }
-  },
-  
 // EDIT 관련
 
   updateProblem: async (problemId, problemText, answer, status, memo) => {
