+++ srcs/problem/problem.sql.js
@@ -1,79 +1,4 @@
 export const sql = {
-  searchProblem: `
-    SELECT 
-      *
-    FROM 
-      problem
-    WHERE 
-      problem_text LIKE ?
-  `,
-
-  searchProblemByFolder: `
-    SELECT 
-      *
-    FROM 
-      problem
-    WHERE 
-      problem_text LIKE ? AND folder_id = ?
-  `,
-
-  checkSubscriptionStatus: `
-    SELECT 
-      status,
-      start_date,
-      end_date
-    FROM 
-      subscription
-    WHERE 
-      user_id = ? AND status = '구독' AND NOW() BETWEEN start_date AND end_date
-  `,
-  findProblemById: `
-    SELECT *
-    FROM 
-      problem
-    WHERE 
-      problem_id = ?
-  `,
-
-  findPhotosByProblemId: `
-    SELECT 
-      photo_url,
-      photo_type
-    FROM 
-      photo
-    WHERE 
-      problem_id = ?
-  `,
-
-  findTypesByProblemId: `
-    SELECT 
-      pt.type_name,
-      pt.type_level
-    FROM 
-      problemtype pt
-    JOIN 
-      problemtypeassignment pta ON pt.type_id = pta.type_id
-    WHERE 
-      pta.problem_id = ?
-  `,
-  // EDIT 관련
-
- // 문제 텍스트 및 정답 업데이트
-  updateProblem: `
-    UPDATE problem
-    SET 
-      problem_text = ?, 
-      answer = ?, 
-      updated_at = NOW()
-    WHERE 
-      problem_id = ?
-  `,
-
-  // 기존 유형 할당 삭제
-  deleteProblemTypeAssignment: `
-    DELETE FROM problemtypeassignment WHERE problem_id = ?
-  `,
-
   updateProblem: `
     UPDATE problem
     SET
