+++ srcs/problem/statistic/statistic.sql.js
